olive/passes/onnx/ryzen_ai/examples/onnx/image_classification/requirements.txt
numpy
onnx>=1.16.0
onnxruntime>=1.20.1,<=1.22.0
onnxruntime_extensions
Pillow
protobuf
rich
torchvision
